optimize e2e restart kubelet
